[Move] 파일구조변경
