+++ api/liveKitApi.ts
@@ -1,4 +1,4 @@
-import { MafiaRoom } from "../_types";
+import { MafiaRoom } from "../types";
 
 export async function getToken({ room, name }: MafiaRoom) {
   const response = await fetch(`/livekit/get-participant-token?room=${room}&username=${name}`, {
