+++ app/(withGlobalLayout)/callback/route.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/app/_utils/supabase/server";
+import { createClient } from "@/utils/supabase/server";
 import { NextResponse } from "next/server";
 
 export async function GET(request: Request) {
