+++ app/(withGlobalLayout)/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../_components/layout/Header";
-import Footer from "../_components/layout/Footer";
+import Header from "../../components/layout/Header";
+import Footer from "../../components/layout/Footer";
 
 const WithGlobalLayout = ({ children }: { children: React.ReactNode }) => {
   return (
