+++ app/(withGlobalLayout)/login/page.tsx
@@ -2,7 +2,7 @@
 
 import { MouseEvent, useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
-import { checkUserLogIn, emailLogIn, oAuthLogIn } from "../../_utils/supabase/authAPI";
+import { checkUserLogIn, emailLogIn, oAuthLogIn } from "../../../utils/supabase/authAPI";
 
 const LogIn = () => {
   const [email, setEmail] = useState("");
