+++ components/layout/Header.tsx
@@ -3,7 +3,7 @@ import Nav from "./Nav";
 import Link from "next/link";
 import Logo from "@/public/images/logo.svg";
 import Image from "next/image";
-import S from "@/app/_style/commons/commons.module.css";
+import S from "@/style/commons/commons.module.css";
 
 const Header = () => {
   return (
