+++ components/layout/Nav.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import React from "react";
-import S from "@/app/_style/commons/commons.module.css";
+import S from "@/style/commons/commons.module.css";
 
 const Nav = () => {
   return (
