+++ style/mainPage/main.module.css
@@ -15,7 +15,7 @@
   display: flex;
   align-items: center;
   height: inherit;
-  background: url("../../../public/images/mafia_visual.png") center center/cover no-repeat;
+  background: url("../../public/images/mafia_visual.png") center center/cover no-repeat;
 }
 
 .roomSectionWrap {
