+++ utils/mafia-algorithm/class/citizenClass.js
