+++ utils/mafia-algorithm/class/doctorClass.js
