+++ utils/mafia-algorithm/class/mafiaClass.js
