+++ utils/mafia-algorithm/class/moderatorClass.js
