+++ utils/mafia-algorithm/class/policeClass.js
