+++ utils/mafia-algorithm/gamePlay.js
