+++ utils/mafia-algorithm/play-variable.js
